dvds

+++ monitor.h
@@ -2,7 +2,7 @@
 #define MONITOR_H
 
 void init_monitor();
-double get_current_angle_degree();
-float get_current_angle();
+float get_current_angle_degree();
+int get_current_angle();
 
 #endif
